Make easier to search for a route.
When searching (typing) for a route ignore accents in characters.
The simplest way to do this would be to convert everything to ASCII and
compare after. However since SQLite is limited in this regard, the
'glob' operator is used instead, which provides a minimal regular
expression support. With that it's possible to "normalize" the text
being queried and extend it with all possible non-ASCII characters.